+++ features/fixtures/simple/Gemfile
@@ -1,3 +1,7 @@
 source 'https://rubygems.org'
 
-gem 'activesupport'
+gem 'activesupport', '7.0.4.3'
+gem 'concurrent-ruby',  '1.2.2'
+gem 'i18n', '1.12.0'
+gem 'minitest', '5.18.0'
+gem 'tzinfo', '2.0.6'
