+++ lib/bom_component.rb
@@ -1,45 +1,2 @@
-
-class BomComponent
-  DEFAULT_TYPE = "library".freeze
-  HASH_ALG = 'SHA-256'.freeze
-
-  def initialize(gem)
-    @name = gem['name']
-    @version = gem['version']
-    @description = gem['description']
-    @hash = gem['hash']
-    @purl = gem['purl']
-    @gem = gem
-  end
-
-  def hash_val
-    component_hash = {
-      "type": DEFAULT_TYPE,
-      "name": @name,
-      "version": @version,
-      "description": @description,
-      "purl": @purl,
-      "hashes": [
-          "alg": HASH_ALG,
-          "content": @hash
-      ]
-    }
-
-    if @gem['license_id']
-      component_hash[:"licenses"] = [
-        "license": {
-          "id": @gem['license_id']
-        }
-      ]
-    elsif @gem['license_name']
-      component_hash[:"licenses"] = [
-        "license": {
-          "name": @gem['license_name']
-        }
-      ]
-    end
-
-    [component_hash]
-
-  end
-end
+require_relative "cyclonedx/ruby"
+require_relative "cyclonedx_deprecated"
