+++ lib/bom_helpers.rb
@@ -22,85 +22,5 @@
 #
 # frozen_string_literal: true
 
-require_relative 'bom_component'
-
-def purl(name, version)
-  "pkg:gem/#{name}@#{version}"
-end
-
-def random_urn_uuid
-  "urn:uuid:#{SecureRandom.uuid}"
-end
-
-def build_bom(gems, format)
-  if format == 'json'
-    build_json_bom(gems)
-  else
-    build_bom_xml(gems)
-  end
-end
-
-def build_json_bom(gems)
-  bom_hash = {
-    "bomFormat": "CycloneDX",
-    "specVersion": "1.1",
-    "serialNumber": random_urn_uuid,
-    "version": 1,
-    "components": []
-  }
-
-  gems.each do |gem|
-    bom_hash[:components] += BomComponent.new(gem).hash_val()
-  end
-
-  JSON.pretty_generate(bom_hash)
-end
-
-def build_bom_xml(gems)
-  builder = Nokogiri::XML::Builder.new(encoding: 'UTF-8') do |xml|
-    attributes = { 'xmlns' => 'http://cyclonedx.org/schema/bom/1.1', 'version' => '1', 'serialNumber' => random_urn_uuid }
-    xml.bom(attributes) do
-      xml.components do
-        gems.each do |gem|
-          xml.component('type' => 'library') do
-            xml.name gem['name']
-            xml.version gem['version']
-            xml.description gem['description']
-            xml.hashes  do
-              xml.hash_ gem['hash'], alg: 'SHA-256'
-            end
-            if gem['license_id']
-              xml.licenses do
-                xml.license do
-                  xml.id gem['license_id']
-                end
-              end
-            elsif gem['license_name']
-              xml.licenses do
-                xml.license do
-                  xml.name gem['license_name']
-                end
-              end
-            end
-            xml.purl gem['purl']
-          end
-        end
-      end
-    end
-  end
-
-  builder.to_xml
-end
-
-def get_gem(name, version)
-  url = "https://rubygems.org/api/v1/versions/#{name}.json"
-  begin
-    RestClient.proxy = ENV['http_proxy']
-    response = RestClient.get(url)
-    body = JSON.parse(response.body)
-    body.select { |item| item['number'] == version.to_s }.first
-  rescue StandardError
-    @logger.warn("#{name} couldn't be fetched")
-    nil
-  end
-end
+require_relative "cyclonedx/ruby"
+require_relative "cyclonedx_deprecated"
