+++ lib/cyclonedx/ruby/version.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Cyclonedx
+  module Ruby
+    module Version
+      VERSION = "1.2.0"
+    end
+    VERSION = Version::VERSION # Make VERSION available in traditional way
+  end
+end
