+++ sig/cyclonedx/ruby.rbs
@@ -1,0 +1,6 @@
+module Cyclonedx
+  module Ruby
+    VERSION: String
+    # See the writing guide of rbs: https://github.com/ruby/rbs#guides
+  end
+end
