+++ spec/cyclonedx/ruby_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe Cyclonedx::Ruby do
+  it "has a version number" do
+    expect(Cyclonedx::Ruby::VERSION).not_to be nil
+  end
+end
